+++ vscp/test/vscpUser/vscp_portable.c
@@ -191,6 +191,23 @@ extern void vscp_portable_provideEvent(vscp_RxMessage const * const msg)
     return;
 }
 
+#if VSCP_CONFIG_BASE_IS_ENABLED( VSCP_CONFIG_ENABLE_SEGMENT_TIME_CALLOUT )
+
+/**
+ * This function is called for every received segment master heartbeat event,
+ * in case it contains a new time since epoch.
+ * 
+ * @param timestamp Unix timestamp
+ */
+extern void vscp_portable_updateTimeSinceEpoch(uint32_t timestamp)
+{
+    vscp_test_portableUpdateTimeSinceEpoch(timestamp);
+
+    return;
+}
+
+#endif  /* VSCP_CONFIG_BASE_IS_ENABLED( VSCP_CONFIG_ENABLE_SEGMENT_TIME_CALLOUT ) */
+
 /*******************************************************************************
     LOCAL FUNCTIONS
 *******************************************************************************/